+++ explauto/sensorimotor_model/forward/nn.py
@@ -30,7 +30,7 @@ class NNForwardModel(ForwardModel):
         @return    predicted y as np.array of float
         """
         _, indexes = self.dataset.nn_x(xq, k = 1)
-        return self.dataset.get_y(indexes[0])
+        return self.dataset.get_y(indexes[0]) # get the sensory effect given index of motor action
         
     def predict_given_context(self, x, c, c_dims):
         """Provide a prediction of x with context c on dimensions c_dims in the output space being S - c_dims
@@ -94,16 +94,16 @@ class NSNNForwardModel(ForwardModel):
 
     def _weights(self, dists, index, sigma_sq):
         w = np.fromiter((gaussian_kernel(d, sigma_sq)
-                         for d in dists), np.float, len(dists))
+                         for d in dists), float, len(dists))
 
         # Weight by timestamp of samples to forget old values
         max_index = max(index)        
         wt = np.fromiter((gaussian_kernel(max_index - idx, self.sigma_t_sq)
-                         for idx in index), np.float, len(dists))
+                         for idx in index), float, len(dists))
         w = w * wt        
         wsum = w.sum()
         if wsum == 0:
             return 1.0/len(dists)*np.ones((len(dists),))
         else:
             eps = wsum * 1e-10 / self.dim_x
-            return np.fromiter((w_i/wsum if w_i > eps else 0.0 for w_i in w), np.float)
+            return np.fromiter((w_i/wsum if w_i > eps else 0.0 for w_i in w), float)
